add tests for modes
- fix few command names
