+++ obdii/__init__.py
@@ -7,11 +7,14 @@ __version__ = "0.1.1a0"
 from logging import NullHandler, getLogger
 
 from .connection import Connection
-from .commands import commands
-from .modes.modeat import at_commands
+from .commands import Commands
+from .modes import at_commands
 
 # We must __init__ .protocols to BaseProtocol.register supported protocols
 from .protocols import *
+
+# Initialize Commands
+commands = Commands()
 
 __all__ = [
     "Connection",
